Address review feedback on #32
- Make init() idempotent in SQLite, Redis, IndexedDB, and LMDB adapters
  to prevent leaking connections on repeated calls
- Fix LMDB clear() to close old database handle before reinitializing
- Fix file-system adapter filename encoding to include % and control
  characters, preventing collisions
- Fix OPFS getMany() to return empty array instead of throwing
  BatchOperationError when no IDs are found
- Remove redundant database.init() call in VectorDB.init() since
  VectorStorage.init() already handles it
- Fix test cleanup in FS, SQLite, Level, and LMDB test files to call
  adapter.destroy() before removing temp directories
- Fix updateVector/updateMetadata in file-system, LMDB, and S3 adapters
  to read data without inflating access counts
- Extract shared serialization utilities (calculateMagnitude,
  vectorDataToJson, jsonToVectorData) into a shared module to eliminate
  duplication across seven adapter files
