+++ src/storage/adapters/lmdb-adapter.ts
@@ -7,6 +7,11 @@ import type {
   StorageAdapter,
   VectorData,
 } from '@/core/types.js';
+import {
+  calculateMagnitude,
+  jsonToVectorData,
+  vectorDataToJson,
+} from './serialization.js';
 
 // ---------------------------------------------------------------------------
 // LMDB types (declared inline because lmdb is an optional peer dependency)
@@ -43,96 +48,6 @@ interface LmdbStorageAdapterOptions {
 }
 
 // ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-function calculateMagnitude(vector: Float32Array): number {
-  let sum = 0;
-  for (let i = 0; i < vector.length; i++) {
-    sum += vector[i]! * vector[i]!;
-  }
-  return Math.sqrt(sum);
-}
-
-// ---------------------------------------------------------------------------
-// Serialization — JSON format
-// ---------------------------------------------------------------------------
-
-interface SerializedVectorData {
-  id: string;
-  vector: number[];
-  metadata?: Record<string, unknown>;
-  magnitude: number;
-  format?: string;
-  normalized?: boolean;
-  timestamp: number;
-  lastAccessed?: number;
-  accessCount?: number;
-  compression?: VectorData['compression'];
-}
-
-function vectorDataToJson(data: VectorData): string {
-  const serialized: SerializedVectorData = {
-    id: data.id,
-    vector: Array.from(data.vector),
-    magnitude: data.magnitude,
-    timestamp: data.timestamp,
-  };
-
-  if (data.metadata !== undefined) {
-    serialized.metadata = data.metadata;
-  }
-  if (data.format !== undefined) {
-    serialized.format = data.format;
-  }
-  if (data.normalized !== undefined) {
-    serialized.normalized = data.normalized;
-  }
-  if (data.lastAccessed !== undefined) {
-    serialized.lastAccessed = data.lastAccessed;
-  }
-  if (data.accessCount !== undefined) {
-    serialized.accessCount = data.accessCount;
-  }
-  if (data.compression !== undefined) {
-    serialized.compression = data.compression;
-  }
-
-  return JSON.stringify(serialized);
-}
-
-function jsonToVectorData(json: string): VectorData {
-  const parsed = JSON.parse(json) as SerializedVectorData;
-  const result: VectorData = {
-    id: parsed.id,
-    vector: new Float32Array(parsed.vector),
-    magnitude: parsed.magnitude,
-    timestamp: parsed.timestamp,
-  };
-
-  if (parsed.metadata !== undefined) {
-    result.metadata = parsed.metadata;
-  }
-  if (parsed.format !== undefined) {
-    result.format = parsed.format;
-  }
-  if (parsed.normalized !== undefined) {
-    result.normalized = parsed.normalized;
-  }
-  if (parsed.lastAccessed !== undefined) {
-    result.lastAccessed = parsed.lastAccessed;
-  }
-  if (parsed.accessCount !== undefined) {
-    result.accessCount = parsed.accessCount;
-  }
-  if (parsed.compression !== undefined) {
-    result.compression = parsed.compression;
-  }
-
-  return result;
-}
-
-// ---------------------------------------------------------------------------
 // LmdbStorageAdapter
 // ---------------------------------------------------------------------------
 
@@ -149,6 +64,11 @@ export class LmdbStorageAdapter implements StorageAdapter {
   // ── Lifecycle ───────────────────────────────────────────────────────────
 
   async init(): Promise<void> {
+    // Make init idempotent: if a database handle already exists, do nothing.
+    if (this.database !== null) {
+      return;
+    }
+
     // lmdb is an optional peer dependency -- use a variable to avoid
     // TypeScript's static module resolution on the import specifier.
     const moduleName = 'lmdb';
@@ -294,8 +214,11 @@ export class LmdbStorageAdapter implements StorageAdapter {
     const database = this.requireDatabase();
     await database.drop();
 
-    // Reinitialize after drop since the database is still open but the
-    // named sub-database has been removed.
+    // Close the old handle and null it out so init() creates a fresh one.
+    await database.close();
+    this.database = null;
+
+    // Reinitialize after drop since the database handle is now stale.
     await this.init();
   }
 
@@ -337,7 +260,7 @@ export class LmdbStorageAdapter implements StorageAdapter {
     vector: Float32Array,
     options?: { updateMagnitude?: boolean; updateTimestamp?: boolean },
   ): Promise<void> {
-    const existing = await this.get(id);
+    const existing = this.readExisting(id);
 
     existing.vector = vector;
 
@@ -358,7 +281,7 @@ export class LmdbStorageAdapter implements StorageAdapter {
     metadata: Record<string, unknown>,
     options?: { merge?: boolean; updateTimestamp?: boolean },
   ): Promise<void> {
-    const existing = await this.get(id);
+    const existing = this.readExisting(id);
 
     if (options?.merge !== false && existing.metadata) {
       existing.metadata = { ...existing.metadata, ...metadata };
@@ -392,7 +315,7 @@ export class LmdbStorageAdapter implements StorageAdapter {
 
     for (const update of updates) {
       try {
-        const existing = await this.get(update.id);
+        const existing = this.readExisting(update.id);
 
         if (update.vector) {
           existing.vector = update.vector;
@@ -423,6 +346,18 @@ export class LmdbStorageAdapter implements StorageAdapter {
   }
 
   // ── Internal helpers ────────────────────────────────────────────────────
+
+  /** Read a vector from the database without updating access tracking. */
+  private readExisting(id: string): VectorData {
+    const database = this.requireDatabase();
+    const json = database.get(id);
+
+    if (json === undefined) {
+      throw new VectorNotFoundError(id);
+    }
+
+    return jsonToVectorData(json);
+  }
 
   private requireDatabase(): LmdbDatabase {
     if (!this.database) {
