+++ src/storage/adapters/redis-adapter.ts
@@ -5,6 +5,11 @@ import type {
   StorageAdapter,
   VectorData,
 } from '@/core/types.js';
+import {
+  calculateMagnitude,
+  jsonToVectorData,
+  vectorDataToJson,
+} from './serialization.js';
 
 // ---------------------------------------------------------------------------
 // Bun.RedisClient types (declared inline since the @types/bun version may
@@ -35,96 +40,6 @@ interface BunRedisClientConstructor {
 interface RedisStorageAdapterOptions {
   url?: string;
   prefix: string;
-}
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-function calculateMagnitude(vector: Float32Array): number {
-  let sum = 0;
-  for (let i = 0; i < vector.length; i++) {
-    sum += vector[i]! * vector[i]!;
-  }
-  return Math.sqrt(sum);
-}
-
-// ---------------------------------------------------------------------------
-// Serialization — JSON format
-// ---------------------------------------------------------------------------
-
-interface SerializedVectorData {
-  id: string;
-  vector: number[];
-  metadata?: Record<string, unknown>;
-  magnitude: number;
-  format?: string;
-  normalized?: boolean;
-  timestamp: number;
-  lastAccessed?: number;
-  accessCount?: number;
-  compression?: VectorData['compression'];
-}
-
-function vectorDataToJson(data: VectorData): string {
-  const serialized: SerializedVectorData = {
-    id: data.id,
-    vector: Array.from(data.vector),
-    magnitude: data.magnitude,
-    timestamp: data.timestamp,
-  };
-
-  if (data.metadata !== undefined) {
-    serialized.metadata = data.metadata;
-  }
-  if (data.format !== undefined) {
-    serialized.format = data.format;
-  }
-  if (data.normalized !== undefined) {
-    serialized.normalized = data.normalized;
-  }
-  if (data.lastAccessed !== undefined) {
-    serialized.lastAccessed = data.lastAccessed;
-  }
-  if (data.accessCount !== undefined) {
-    serialized.accessCount = data.accessCount;
-  }
-  if (data.compression !== undefined) {
-    serialized.compression = data.compression;
-  }
-
-  return JSON.stringify(serialized);
-}
-
-function jsonToVectorData(json: string): VectorData {
-  const parsed = JSON.parse(json) as SerializedVectorData;
-  const result: VectorData = {
-    id: parsed.id,
-    vector: new Float32Array(parsed.vector),
-    magnitude: parsed.magnitude,
-    timestamp: parsed.timestamp,
-  };
-
-  if (parsed.metadata !== undefined) {
-    result.metadata = parsed.metadata;
-  }
-  if (parsed.format !== undefined) {
-    result.format = parsed.format;
-  }
-  if (parsed.normalized !== undefined) {
-    result.normalized = parsed.normalized;
-  }
-  if (parsed.lastAccessed !== undefined) {
-    result.lastAccessed = parsed.lastAccessed;
-  }
-  if (parsed.accessCount !== undefined) {
-    result.accessCount = parsed.accessCount;
-  }
-  if (parsed.compression !== undefined) {
-    result.compression = parsed.compression;
-  }
-
-  return result;
 }
 
 // ---------------------------------------------------------------------------
@@ -161,6 +76,11 @@ export class RedisStorageAdapter implements StorageAdapter {
   // ── Lifecycle ───────────────────────────────────────────────────────────
 
   async init(): Promise<void> {
+    // Make init idempotent: if a client already exists, do nothing.
+    if (this.client) {
+      return;
+    }
+
     if (typeof Bun === 'undefined' || !Bun.RedisClient) {
       throw new Error(
         'RedisStorageAdapter requires the Bun runtime with built-in Redis support',
